requirements.txt
# LightClaw - Minimal AI Agent
# Core dependencies only

python-telegram-bot~=21.0
openai~=1.0
anthropic~=0.40
google-generativeai~=0.8
python-dotenv~=1.0
numpy~=1.26
